LayoutTests/fast/js/array-prototype-properties-expected.txt

This is a test case for bug 64679.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Array.prototype.toString.call(undefined) threw exception TypeError: Type error.
PASS Array.prototype.toLocaleString.call(undefined) threw exception TypeError: Type error.
FAIL Array.prototype.concat.call(undefined, []) should throw an exception. Was [object DOMWindow].
FAIL Array.prototype.join.call(undefined, []) should throw an exception. Was .
FAIL Array.prototype.pop.call(undefined) should throw an exception. Was undefined.
FAIL Array.prototype.push.call(undefined, {}) should throw an exception. Was 1.
FAIL Array.prototype.reverse.call(undefined) should throw an exception. Was [object DOMWindow].
FAIL Array.prototype.shift.call(undefined) should throw an exception. Was [object Object].
FAIL Array.prototype.slice.call(undefined, 0, 1) should throw an exception. Was .
FAIL Array.prototype.sort.call(undefined) should throw an exception. Was [object DOMWindow].
FAIL Array.prototype.splice.call(undefined, 0, 1) should throw an exception. Was .
FAIL Array.prototype.unshift.call(undefined, {}) should throw an exception. Was 1.
FAIL Array.prototype.every.call(undefined, toString) should throw an exception. Was true.
FAIL Array.prototype.forEach.call(undefined, toString) should throw an exception. Was undefined.
FAIL Array.prototype.some.call(undefined, toString) should throw an exception. Was true.
FAIL Array.prototype.indexOf.call(undefined, 0) should throw an exception. Was -1.
PASS Array.prototype.indlastIndexOfexOf.call(undefined, 0) threw exception TypeError: 'undefined' is not an object (evaluating 'Array.prototype.indlastIndexOfexOf.call').
FAIL Array.prototype.filter.call(undefined, toString) should throw an exception. Was [object Object].
FAIL Array.prototype.reduce.call(undefined, toString) should throw an exception. Was [object Object].
FAIL Array.prototype.reduceRight.call(undefined, toString) should throw an exception. Was [object Object].
FAIL Array.prototype.map.call(undefined, toString) should throw an exception. Was [object DOMWindow].
PASS successfullyParsed is true

TEST COMPLETE
